server

+++ Models/Services/CartService.js
@@ -5,6 +5,7 @@ import Op from "sequelize";
 export const createCartService = async(quantity, BookId, customerId) =>{
     try {
         const book = await db.book.findByPk(BookId)
+        if(!quantity || quantity == 0) return createError(400, 'Vui lòng chọn số lượng!')
         if(!book) return createError(400, 'Không tìm thấy sach!')
         const checkCart = await db.cart.findOne({
             where :{
@@ -43,15 +44,12 @@ export const createCartService = async(quantity, BookId, customerId) =>{
 
 export const deleteCartService = async(id, customerId) =>{
     try {
-        console.log(id)
         const cart = await db.cart.findByPk(id);
         if(!cart) return createError(400, 'Không tìm thấy giỏ hàng!')
+        if(cart.customerId !== customerId) return createError(400, 'Bạn không thể xó giỏ hầng người khác')
         const delete_cart = await db.cart.destroy({
             where : {
-                [Op.and] : [
-                    {id: id}, 
-                    {customerId: customerId}
-                ]
+                id
             }
         })
         if(delete_cart == 0) return createError(400, 'Xoá giỏ hàng không thành công!')
@@ -61,12 +59,11 @@ export const deleteCartService = async(id, customerId) =>{
     }
 }
 
-export const getCartByIdService = async(id) =>{
+export const getCartByIdService = async(id, customerId) =>{
     try {
-        const cart = await db.cart.findOne({
-            where: {id}
-        })
-        if(!cart) return createError(400, 'Tìm kiếm không thành công!')
+        const cart = await db.cart.findByPk(id);
+        if(!cart) return createError(400, 'Không tìm thấy giỏ hàng!')
+        if(cart.customerId !== customerId) return createError(400, 'Bạn không thể xem giỏ hầng người khác')
         return cart;
     } catch (error) {
         return error;
